fix(fmt): lint

+++ src/mocks/ERC20Mock.sol
@@ -6,7 +6,7 @@ import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 contract ERC20Mock is ERC20 {
     constructor(string memory name_, string memory symbol_) ERC20(name_, symbol_) {}
 
-    function mint(address _to, uint _amount) public {
+    function mint(address _to, uint256 _amount) public {
         _mint(_to, _amount);
     }
 }
