+++ src/token/oft/extension/ProxyOFT.sol
@@ -14,18 +14,27 @@ contract ProxyOFT is OFTCore {
         token = IERC20(_proxyToken);
     }
 
-    function circulatingSupply() public view virtual override returns (uint) {
+    function circulatingSupply() public view virtual override returns (uint256) {
         unchecked {
             return token.totalSupply() - token.balanceOf(address(this));
         }
     }
 
-    function _debitFrom(address _from, uint16, bytes memory, uint _amount) internal virtual override {
+    function _debitFrom(
+        address _from,
+        uint16,
+        bytes memory,
+        uint256 _amount
+    ) internal virtual override {
         require(_from == _msgSender(), "ProxyOFT: owner is not send caller");
         token.safeTransferFrom(_from, address(this), _amount);
     }
 
-    function _creditTo(uint16, address _toAddress, uint _amount) internal virtual override {
+    function _creditTo(
+        uint16,
+        address _toAddress,
+        uint256 _amount
+    ) internal virtual override {
         token.safeTransfer(_toAddress, _amount);
     }
 }
