+++ src/token/onft/IONFT1155Core.sol
@@ -8,10 +8,38 @@ import "@openzeppelin/contracts/utils/introspection/IERC165.sol";
  * @dev Interface of the ONFT Core standard
  */
 interface IONFT1155Core is IERC165 {
-    event SendToChain(address indexed _sender, uint16 indexed _dstChainId, bytes indexed _toAddress, uint _tokenId, uint _amount, uint64 _nonce);
-    event SendBatchToChain(address indexed _sender, uint16 indexed _dstChainId, bytes indexed _toAddress, uint[] _tokenIds, uint[] _amounts, uint64 _nonce);
-    event ReceiveFromChain(uint16 indexed _srcChainId, bytes indexed _srcAddress, address indexed _toAddress, uint _tokenId, uint _amount, uint64 _nonce);
-    event ReceiveBatchFromChain(uint16 indexed _srcChainId, bytes indexed _srcAddress, address indexed _toAddress, uint[] _tokenIds, uint[] _amounts, uint64 _nonce);
+    event SendToChain(
+        address indexed _sender,
+        uint16 indexed _dstChainId,
+        bytes indexed _toAddress,
+        uint256 _tokenId,
+        uint256 _amount,
+        uint64 _nonce
+    );
+    event SendBatchToChain(
+        address indexed _sender,
+        uint16 indexed _dstChainId,
+        bytes indexed _toAddress,
+        uint256[] _tokenIds,
+        uint256[] _amounts,
+        uint64 _nonce
+    );
+    event ReceiveFromChain(
+        uint16 indexed _srcChainId,
+        bytes indexed _srcAddress,
+        address indexed _toAddress,
+        uint256 _tokenId,
+        uint256 _amount,
+        uint64 _nonce
+    );
+    event ReceiveBatchFromChain(
+        uint16 indexed _srcChainId,
+        bytes indexed _srcAddress,
+        address indexed _toAddress,
+        uint256[] _tokenIds,
+        uint256[] _amounts,
+        uint64 _nonce
+    );
 
     // _from - address where tokens should be deducted from on behalf of
     // _dstChainId - L0 defined chain id to send tokens too
@@ -21,7 +49,16 @@ interface IONFT1155Core is IERC165 {
     // _refundAddress - address on src that will receive refund for any overpayment of L0 fees
     // _zroPaymentAddress - if paying in zro, pass the address to use. using 0x0 indicates not paying fees in zro
     // _adapterParams - flexible bytes array to indicate messaging adapter services in L0
-    function sendFrom(address _from, uint16 _dstChainId, bytes calldata _toAddress, uint _tokenId, uint _amount, address payable _refundAddress, address _zroPaymentAddress, bytes calldata _adapterParams) external payable;
+    function sendFrom(
+        address _from,
+        uint16 _dstChainId,
+        bytes calldata _toAddress,
+        uint256 _tokenId,
+        uint256 _amount,
+        address payable _refundAddress,
+        address _zroPaymentAddress,
+        bytes calldata _adapterParams
+    ) external payable;
 
     // _from - address where tokens should be deducted from on behalf of
     // _dstChainId - L0 defined chain id to send tokens too
@@ -31,7 +68,16 @@ interface IONFT1155Core is IERC165 {
     // _refundAddress - address on src that will receive refund for any overpayment of L0 fees
     // _zroPaymentAddress - if paying in zro, pass the address to use. using 0x0 indicates not paying fees in zro
     // _adapterParams - flexible bytes array to indicate messaging adapter services in L0
-    function sendBatchFrom(address _from, uint16 _dstChainId, bytes calldata _toAddress, uint[] calldata _tokenIds, uint[] calldata _amounts, address payable _refundAddress, address _zroPaymentAddress, bytes calldata _adapterParams) external payable;
+    function sendBatchFrom(
+        address _from,
+        uint16 _dstChainId,
+        bytes calldata _toAddress,
+        uint256[] calldata _tokenIds,
+        uint256[] calldata _amounts,
+        address payable _refundAddress,
+        address _zroPaymentAddress,
+        bytes calldata _adapterParams
+    ) external payable;
 
     // _dstChainId - L0 defined chain id to send tokens too
     // _toAddress - dynamic bytes array which contains the address to whom you are sending tokens to on the dstChain
@@ -39,7 +85,14 @@ interface IONFT1155Core is IERC165 {
     // _amount - amount of the tokens to transfer
     // _useZro - indicates to use zro to pay L0 fees
     // _adapterParams - flexible bytes array to indicate messaging adapter services in L0
-    function estimateSendFee(uint16 _dstChainId, bytes calldata _toAddress, uint _tokenId, uint _amount, bool _useZro, bytes calldata _adapterParams) external view returns (uint nativeFee, uint zroFee);
+    function estimateSendFee(
+        uint16 _dstChainId,
+        bytes calldata _toAddress,
+        uint256 _tokenId,
+        uint256 _amount,
+        bool _useZro,
+        bytes calldata _adapterParams
+    ) external view returns (uint256 nativeFee, uint256 zroFee);
 
     // _dstChainId - L0 defined chain id to send tokens too
     // _toAddress - dynamic bytes array which contains the address to whom you are sending tokens to on the dstChain
@@ -47,5 +100,12 @@ interface IONFT1155Core is IERC165 {
     // _amounts - amounts of the tokens to transfer
     // _useZro - indicates to use zro to pay L0 fees
     // _adapterParams - flexible bytes array to indicate messaging adapter services in L0
-    function estimateSendBatchFee(uint16 _dstChainId, bytes calldata _toAddress, uint[] calldata _tokenIds, uint[] calldata _amounts, bool _useZro, bytes calldata _adapterParams) external view returns (uint nativeFee, uint zroFee);
+    function estimateSendBatchFee(
+        uint16 _dstChainId,
+        bytes calldata _toAddress,
+        uint256[] calldata _tokenIds,
+        uint256[] calldata _amounts,
+        bool _useZro,
+        bytes calldata _adapterParams
+    ) external view returns (uint256 nativeFee, uint256 zroFee);
 }
